requirements.txt
geopy
matplotlib
Pillow
Requests
tkcalendar
yfinance
